--- a/src/core/mesh/ext/triangle/src/CMakeLists.txt
+++ b/src/core/mesh/ext/triangle/src/CMakeLists.txt
@@ -1,1 +1,0 @@
-target_sources(triangle PRIVATE triangle.cpp)