--- a/src/core/mesh/src/triangulate.hpp
+++ b/src/core/mesh/src/triangulate.hpp
@@ -1,9 +1,6 @@
-// Wrapper around Triangle meshing library
-//   - https://www.cs.cmu.edu/~quake/triangle.html
-//   - takes a set of boundary points & segments connecting them
-//   - also a set of interior points for each compartment
+// Wrapper around a meshing library
+//   - takes boundaries, interior point & max area for each compartment
 //   - generates mesh with triangles labelled according to compartment
-//   - replaces triangles not assigned a compartment with holes and re-meshes
 //   - provides resulting vertices & triangles
 
 #pragma once
@@ -14,35 +11,13 @@
 #include <cstddef>
 #include <string>
 #include <vector>
+#include "boundary.hpp"
 
 namespace sme {
 
 namespace mesh {
 
-class Boundary;
-
-struct TriangulateBoundarySegment {
-  std::size_t start;
-  std::size_t end;
-};
-
-using TriangulateBoundarySegments = std::vector<TriangulateBoundarySegment>;
 using TriangulateTriangleIndex = std::array<std::size_t, 3>;
-
-struct TriangulateCompartment {
-  std::vector<QPointF> interiorPoints;
-  double maxTriangleArea;
-};
-
-struct TriangulateBoundaries {
-  TriangulateBoundaries();
-  TriangulateBoundaries(const std::vector<Boundary> &inputBoundaries,
-                        const std::vector<std::vector<QPointF>> &interiorPoints,
-                        const std::vector<std::size_t> &maxTriangleAreas);
-  std::vector<QPointF> vertices;
-  std::vector<TriangulateBoundarySegments> boundaries;
-  std::vector<TriangulateCompartment> compartments;
-};
 
 class Triangulate {
 private:
@@ -50,7 +25,9 @@ private:
   std::vector<std::vector<TriangulateTriangleIndex>> triangleIndices;
 
 public:
-  explicit Triangulate(const TriangulateBoundaries &boundaries);
+  explicit Triangulate(const std::vector<Boundary> &inputBoundaries,
+                       const std::vector<std::vector<QPointF>> &interiorPoints,
+                       const std::vector<std::size_t> &maxTriangleAreas);
   [[nodiscard]] const std::vector<QPointF> &getPoints() const;
   [[nodiscard]] const std::vector<std::vector<TriangulateTriangleIndex>> &
   getTriangleIndices() const;