… .github/workflows/wheels.yml
… README.md
… benchmark/geometry.png
… benchmark/mesh.png
… benchmark/model.png
… benchmark/simulate.png
… ci/windows-gui.sh
… sme/README.md
… src/core/CMakeLists.txt
… src/core/mesh/CMakeLists.txt
… src/core/mesh/ext/triangle/CMakeLists.txt
… src/core/mesh/ext/triangle/inc/triangle.hpp
… src/core/mesh/ext/triangle/src/triangle.cpp
… src/core/mesh/src/CMakeLists.txt
… src/core/mesh/src/triangulate_bench.cpp
… src/core/mesh/src/triangulate_common.cpp
… src/core/mesh/src/triangulate_t.cpp
… src/gui/dialogs/dialogabout.cpp