- internal mesh::Triangulate interface is unchanged
- max triangle area constraints are converted to max triangle edge length constraints
- a CDT is first constructed from the PLSG
- then compartments are meshed in ascending order of max edge length
- compartments then identified starting from interior point(s)
  - identify triangle corresponding to an interior point
  - recursively add neighbouring triangles until a boundary edge is encountered
- resolves #421