@lkeegan 2021-02-09
meshing with CGAL is 3-4x slower than Triangle:

```
mesh_Mesh<ABtoC>                                                 1,20 ms -> 3,19 ms
mesh_Mesh<VerySimpleModel>                                       1,46 ms -> 3,53 ms
mesh_Mesh<LiverCells>                                            8,32 ms -> 34,4 ms
```

which translates to a regression of ~25% in the GUI meshing performance (since the most expensive part of this is not the meshing, but generating the QImage of the mesh, although this could probably be done faster using opencv):

```
mesh_Mesh_getMeshImages<ABtoC>                                   5,37 ms -> 6,93 ms
mesh_Mesh_getMeshImages<VerySimpleModel>                         8,03 ms -> 10,1 ms
mesh_Mesh_getMeshImages<LiverCells>                              54,2 ms -> 68,2 ms
```